Allow providing an app-db instance when subscribing and dispatching
This adds an extra arity-2 variant of subscribe, dispatch,
dispatch-sync that accepts a "frame", which is any reagent atom the
user wants to operate on.

The frame is passed around as metadata on the event or query vector.
For events, it is injected into with a new coeffect :frame, and the
:db coeffect has changed to look for the :frame coeffect.

The :dispatch-* fx will now accept an optional second frame argument
that :do-fx interceptor will provide.

Subscribe has a new arity to accept a frame as the first argument.
All input signals will operate on the provided frame, no mixing and
matching. Subscription reactions are cached using the frame-id
provided by `reagent-id`, in addition to the usual query-v and dynv
values.

Open questions:
1. Should users be able to dispatch from an event operating on a
provided frame to the greater app-db instance? If so, there will
probably need to be a new set of `:dispatch` fx variants.
In my own use of re-frame I've found this to be something that is
wanted. A user may have a component that accepts an event vector that
uses app-db, but the component itself uses it's own separate frame.
That component will later dispatch to the "gloabl" app-db event it was
given from an event operating on the local frame.

2. What happens when a frame is destoryed? Consider when a reagent
state-atom for a component is being used for state. That component is
destroyed, and the frame goes with it. Some cache may stick
around. Needs testing.

3. Tracing: I don't use the tracing very much. There is one instace in
events/handle where i substituted frame for app-db in a tag called
:app-db-before and :app-db-after. Probably wrong to do that.