new reg-fx variant where handler is also provided the current frame.
:dispatch-* variants, :fx, and :db effects are now registered with
reg-frame-fx.

Instead of keeping a special list of what fx need the frame, I defined
a new handler type that the registrar can accept, :frame-fx, and new
functions to support registering them.

Now the do-fx interceptor and :fx effect will check the registry for
both kinds of effects and provide the frame where appropriate.
This logic was very repetitive, so I also broke out a new private
function `do-fx-without-db` to help the do-fx interceptor and :fx
effect. The console warning in it depends on do-fx removing the db
effect.

Maybe this is not something that should be exposed to the outside
world, as the frame atom is probably best dealt with in events
themselves. Even if it's not going to be public, this implementation
felt a lot less smelly then keeping a special list of :fx handlers
that need a frame to do their work.