Avoid caching a subscription when outside of a reactive context.
This avoids a known memory leak in re-frame where on-dispose may never
be run on a cached subscription, leaving it stuck in the subscription
cache indefinitely.

Open questions:
Should `(trace/merge-trace! {:tags {:reaction (reagent-id r)}})`
always be called in the `cache-and-return` function, even outside a
reactive-context?
What should `reactive-context?` return from interop.clj? From what I
can tell, reagent does not provide a mechanism for checking this in
the jvm.